3624 - Upgrade rails to current version

+++ db/migrate/20161018175439_retire_image_sets.rb
@@ -1,11 +1,10 @@
 class RetireImageSets < ActiveRecord::Migration[5.0]
-  def change
-    if table_exists?(:image_sets)
-      drop_table(:image_sets)
-    end
-    if table_exists?(:titled_images)
-      drop_table(:titled_images)
-    end
 
+  def change
+    drop_table(:image_sets) if table_exists?(:image_sets)
+    return unless table_exists?(:titled_images)
+
+    drop_table(:titled_images)
   end
+
 end
